Fixing tests
Move tests that require torch from test_spark.py to test_spark_torch.py
