+++ test/integration/test_spark.py
@@ -31,16 +31,12 @@ import warnings
 from distutils.version import LooseVersion
 
 import mock
-import torch
-
-import pyspark
 
 from pyspark.ml.linalg import DenseVector, SparseVector, VectorUDT
 from pyspark.sql.types import ArrayType, BooleanType, DoubleType, FloatType, IntegerType, \
     NullType, StructField, StructType
 
 import horovod.spark
-import horovod.torch as hvd
 
 from horovod.common.util import gloo_built, mpi_built
 from horovod.runner.common.util import codec, secret, safe_shell_exec, timeout
@@ -51,7 +47,7 @@ from horovod.spark.common import constants, util
 from horovod.spark.common.store import DBFSLocalStore, HDFSStore, LocalStore, Store
 from horovod.spark.driver.host_discovery import SparkDriverHostDiscovery
 from horovod.spark.driver.rsh import rsh
-from horovod.spark.task import get_available_devices, gloo_exec_fn, mpirun_exec_fn
+from horovod.spark.task import gloo_exec_fn, mpirun_exec_fn
 from horovod.spark.task.task_service import SparkTaskClient
 from horovod.spark.runner import _task_fn
 
@@ -70,20 +66,6 @@ if platform.system() == 'Darwin':
 
 def fn(result=0):
     return result
-
-
-@spawn
-def run_get_available_devices():
-    # Run this test in an isolated "spawned" environment because creating a local-cluster
-    # leads to errors that cause downstream tests to stall:
-    # https://issues.apache.org/jira/browse/SPARK-31922
-    def fn():
-        hvd.init()
-        devices = get_available_devices()
-        return devices, hvd.local_rank()
-
-    with spark_session('test_get_available_devices', gpus=2):
-        return horovod.spark.run(fn, env={'PATH': os.environ.get('PATH')}, verbose=0)
 
 
 class SparkTests(unittest.TestCase):
@@ -341,60 +323,6 @@ class SparkTests(unittest.TestCase):
         self.assertEqual(dict([(index, 123) for index in range(tasks)]), results)
 
     """
-    Test that horovod.spark.run works properly in a simple setup using MPI.
-    """
-    def test_happy_run_with_mpi(self):
-        if not (mpi_built() and is_open_mpi()):
-            self.skipTest("Open MPI is not available")
-
-        self.do_test_happy_run(use_mpi=True, use_gloo=False)
-
-    """
-    Test that horovod.spark.run works properly in a simple setup using Gloo.
-    """
-    def test_happy_run_with_gloo(self):
-        if not gloo_built():
-            self.skipTest("Gloo is not available")
-
-        self.do_test_happy_run(use_mpi=False, use_gloo=True)
-
-    """
-    Actually tests that horovod.spark.run works properly in a simple setup.
-    """
-    def do_test_happy_run(self, use_mpi, use_gloo):
-        def fn():
-            hvd.init()
-            res = hvd.allgather(torch.tensor([hvd.rank()])).tolist()
-            return res, hvd.rank()
-
-        with spark_session('test_happy_run'):
-            with is_built(gloo_is_built=use_gloo, mpi_is_built=use_mpi):
-                res = horovod.spark.run(fn, start_timeout=10,
-                                        use_mpi=use_mpi, use_gloo=use_gloo,
-                                        verbose=2)
-                self.assertListEqual([([0, 1], 0), ([0, 1], 1)], res)
-
-    """
-    Test that horovod.spark.run_elastic works properly in a simple setup.
-    """
-    def test_happy_run_elastic(self):
-        if not gloo_built():
-            self.skipTest("Gloo is not available")
-
-        def fn():
-            # training function does not use ObjectState and @hvd.elastic.run
-            # only testing distribution of state-less training function here
-            # see test_spark_torch.py for testing that
-            hvd.init()
-            res = hvd.allgather(torch.tensor([hvd.rank()])).tolist()
-            return res, hvd.rank()
-
-        with spark_session('test_happy_run_elastic'):
-            res = horovod.spark.run_elastic(fn, num_proc=2, min_np=2, max_np=2,
-                                            start_timeout=10, verbose=2)
-            self.assertListEqual([([0, 1], 0), ([0, 1], 1)], res)
-
-    """
     Test that horovod.spark.run times out when it does not start up fast enough using MPI.
     """
     def test_timeout_with_mpi(self):
@@ -583,7 +511,7 @@ class SparkTests(unittest.TestCase):
         def mpi_impl_flags(tcp, env=None):
             return ["--mock-mpi-impl-flags"], ["--mock-mpi-binding-args"], None
 
-        def gloo_exec_command_fn(driver_addresses, key, settings, env):
+        def gloo_exec_command_fn(driver, key, settings, env, stdout, stderr, prefix_output_with_timestamp):
             def _exec_command(command, alloc_info, event):
                 return 1, alloc_info.rank
             return _exec_command
@@ -756,7 +684,7 @@ class SparkTests(unittest.TestCase):
         num_proc = cores if num_proc is None else num_proc
         self.assertEqual(expected_np, num_proc)
         self.assertEqual(1, gloo_exec_command_fn.call_count)
-        _, _, _, call_env = gloo_exec_command_fn.call_args[0]
+        _, _, _, call_env, _, _, _ = gloo_exec_command_fn.call_args[0]
         self.assertEqual(env or {}, call_env)
         self.assertEqual({}, gloo_exec_command_fn.call_args[1])
         self.assertEqual(num_proc, exec_command.call_count)
@@ -1658,12 +1586,6 @@ class SparkTests(unittest.TestCase):
                 file = os.path.sep.join([d, 'command_executed'])
                 client.abort_command()
                 self.do_test_spark_task_service_executes_command(client, file)
-
-    @pytest.mark.skipif(LooseVersion(pyspark.__version__) < LooseVersion('3.0.0'),
-                        reason='get_available_devices only supported in Spark 3.0 and above')
-    def test_get_available_devices(self):
-        res = run_get_available_devices()
-        self.assertListEqual([(['1'], 0), (['0'], 1)], res)
 
     def test_to_list(self):
         none_output = util.to_list(None, 1)
